requirements.txt
discord.py>=2.3.0
openai>=1.0.0
anthropic>=0.25.0
google-genai>=1.0.0
groq>=0.9.0
python-dotenv>=1.0.0
tenacity>=8.0.0
